Change return type of floormanual to long

+++ soal1.c
@@ -12,7 +12,7 @@
  #include <string.h>
 
 
- int floormanual(double x) {
+ long floormanual(double x) {
     int i = (int)x; // i = integralnya saja dari x
     if (x < 0 && x != i) {
         return i - 1; // situasi negatif
